feat: add order endpoint

+++ backend/food_fast/src/main/java/com/paulotech/food_fast/model/Order.java
@@ -40,5 +40,5 @@ public class Order {
 
     private int totalItems;
 
-    private int totalPrice;
+    private Long totalPrice;
 }
